rename wave_schemas to wave_config, clean up stale names
wave_schemas.rs → wave_config.rs. parse_schema_stimulus → parse_stimulus.
schema_stimulus → config_stimulus. No schema concept remains.
