+++ rust/loopflow/src/lfd/http/routes/mod.rs
@@ -3,8 +3,8 @@ pub mod hooks;
 pub mod repos;
 pub mod sessions;
 pub mod system;
+pub mod wave_config;
 pub mod wave_runs;
-pub mod wave_schemas;
 pub mod waves;
 pub mod worktrees;
 pub mod ws;
